Made my first component
